Implement telemetry
Fixes #29.

* Opted to keep the telemetry calls in the redux middleware, and added
some actions that don't affect the redux store, but are only used by
the telemetry middleware. Seemed cleaner than giving various components
and containers knowledge of the telemetry library.

* Refactored the telemetry methods to use an options object instead of
positional args. In our new metrics, value is optional and sometimes
null, and extra is optional and sometimes null; and it seemed that
getting the positional order right could be bug-prone.

* Added embedded telemetry API experiment to work around recordEvent
bug in Firefox.

* Added an onReveal method to trigger an action, so we can count the
password reveal clicks via telemetry.

* Removed the `reconnecting_sync` event, because we don't yet show
any UI in the FxA menu in the error case.

* Set `record_on_release` to `true` for everything, because our users
are opting in by installing the addon (or clicking a button to enable
it).

* Added telemetry event info to pull request template.